Add an error limit
Add a hard error limit, so that we back off and retry if the
limit is exceeded. The hope is that this will completely solve
the problem of occasional spikes forming, which seem to form
when an unfortunate error is then magnified in subsequent steps.

(All the reductions of INITH that were supposed to solve this problem
made no real difference in the end, since the underlying hypothesis –
that the spikes were caused by large jumps at high-gradient regions
in the early stages – appears to be mistaken.)